avoid include of XBDateTime.h in headers
