+++ xbmc/platform/darwin/DarwinUtils.mm
@@ -29,6 +29,7 @@
 
 #include <mutex>
 
+#include "PlatformDefs.h"
 
 // platform strings are based on http://theiphonewiki.com/wiki/Models
 const char* CDarwinUtils::getIosPlatformString(void)
