+++ xbmc/platform/linux/FDEventMonitor.cpp
@@ -18,6 +18,8 @@
 #include <sys/eventfd.h>
 #include <unistd.h>
 
+#include "PlatformDefs.h"
+
 CFDEventMonitor::CFDEventMonitor() :
   CThread("FDEventMonitor")
 {
