+++ xbmc/utils/DumbBufferObject.cpp
@@ -19,6 +19,8 @@
 #include <sys/mman.h>
 #include <unistd.h>
 
+#include "PlatformDefs.h"
+
 using namespace KODI::WINDOWING::GBM;
 
 std::unique_ptr<CBufferObject> CDumbBufferObject::Create()
