+++ xbmc/utils/StringUtils.h
@@ -31,7 +31,7 @@
 #undef FMT_DEPRECATED
 #define FMT_DEPRECATED
 #endif
-#include "XBDateTime.h"
+#include "utils/TimeFormat.h"
 #include "utils/params_check_macros.h"
 
 #include <fmt/format.h>
