+++ xbmc/utils/XMLUtils.cpp
@@ -7,8 +7,10 @@
  */
 
 #include "XMLUtils.h"
-#include "URL.h"
+
 #include "StringUtils.h"
+#include "URL.h"
+#include "XBDateTime.h"
 
 bool XMLUtils::GetHex(const TiXmlNode* pRootNode, const char* strTag, uint32_t& hexValue)
 {
