+++ xbmc/utils/test/TestArchive.cpp
@@ -10,11 +10,11 @@
 #  include <windows.h>
 #endif
 
+#include "filesystem/File.h"
+#include "test/TestUtils.h"
 #include "utils/Archive.h"
 #include "utils/Variant.h"
-#include "filesystem/File.h"
-
-#include "test/TestUtils.h"
+#include "utils/XTimeUtils.h"
 
 #include <gtest/gtest.h>
