+++ xbmc/windowing/android/WinSystemAndroidGLESContext.cpp
@@ -22,6 +22,8 @@
 #include <EGL/eglext.h>
 #include <unistd.h>
 
+#include "PlatformDefs.h"
+
 void CWinSystemAndroidGLESContext::Register()
 {
   KODI::WINDOWING::CWindowSystemFactory::RegisterWindowSystem(CreateWinSystem);
